Fix 4xx inverted, set some defaults. (#1)
* 4xx % was accidentally inverted.
* Default all y axis mins to 0 since nothing can go -ve
* Use response_code_class!="5" for Success Rate to avoid null = 100%.
* Use null = 0 for 4xx and success rate graph. Envoy doesn't have counters
for e.g. response_code_class=4 if there haven't been any of that class yet,
so we end up with null data if no requests result in 4xx (or !5xx for
success rate.) ie if they should be 0%. Downside is that no requests at
all also leads to a 0% 4xx/success rate, but I'll accept that for now.